MES-607: Refactor of back to office module for category separation
