+++ src/pages/back-to-office/back-to-office.module.ts
@@ -1,18 +1,9 @@
 import { NgModule } from '@angular/core';
-import { IonicPageModule } from 'ionic-angular';
-import { BackToOfficePage } from './back-to-office';
-import { EffectsModule } from '@ngrx/effects';
-import { BackToOfficeAnalyticsEffects } from './back-to-office.analytics.effects';
-import { ComponentsModule } from '../../components/common/common-components.module';
+import { BackToOfficeCatBPageModule } from './cat-b/back-to-office.cat-b.page.module';
 
 @NgModule({
-  declarations: [
-    BackToOfficePage,
-  ],
   imports: [
-    IonicPageModule.forChild(BackToOfficePage),
-    EffectsModule.forFeature([BackToOfficeAnalyticsEffects]),
-    ComponentsModule,
+    BackToOfficeCatBPageModule,
   ],
 })
 export class BackToOfficePageModule {}
