+++ src/pages/back-to-office/cat-b/back-to-office.cat-b.page.html
@@ -4,7 +4,7 @@
 </ion-header>
 <ion-content no-bounce id="back-to-office-page">
     <div class="container">
-        <img id='dvsa-logo' src='../../assets/imgs/back-to-office/logo.png' />
+        <img id='dvsa-logo' src='../../../assets/imgs/back-to-office/logo.png' />
     </div>
     <div class="container">
         <button ion-button class="mes-neutral-button office-button top-button" navPush="OfficePage" id="continue-to-write-up">
