+++ src/pages/back-to-office/cat-b/back-to-office.cat-b.page.module.ts
@@ -1,0 +1,22 @@
+import { NgModule } from '@angular/core';
+import { IonicPageModule } from 'ionic-angular';
+import { ComponentsModule } from '../../../components/common/common-components.module';
+import { BackToOfficeCatBPage } from './back-to-office.cat-b.page';
+import { TranslateModule } from 'ng2-translate';
+import { EffectsModule } from '@ngrx/effects';
+import { BackToOfficeAnalyticsEffects } from '../back-to-office.analytics.effects';
+
+@NgModule({
+  declarations: [
+    BackToOfficeCatBPage,
+  ],
+  imports: [
+    IonicPageModule.forChild(BackToOfficeCatBPage),
+    EffectsModule.forFeature([
+      BackToOfficeAnalyticsEffects,
+    ]),
+    ComponentsModule,
+    TranslateModule,
+  ],
+})
+export class BackToOfficeCatBPageModule { }
