+++ src/pages/back-to-office/cat-b/back-to-office.cat-b.page.scss
@@ -1,5 +1,5 @@
-page-back-to-office {
-  
+back-to-office-cat-b-page {
+
   div {
     background-color: map-get($colors-gds, gds-dark-green-2);
   }
@@ -27,5 +27,5 @@ page-back-to-office {
     margin-top: 64px;
   }
 
-  
+
 }
